fix: 🔨 Saving embed data correctly, apply textarena blockquote via css class
