+++ src/plugins/blockquotePlugin.ts
@@ -8,7 +8,7 @@ import utils from '../utils';
 const defaultOptions: DefaultPluginOptions = {
   name: 'blockquote',
   tag: 'BLOCKQUOTE',
-  attributes: {},
+  attributes: { class: 'textarena-blockquote' },
   title: 'Blockquote',
   icon: `<svg width="16px" height="10px" viewBox="0 0 16 10" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
   <g id="Icons" stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
